# ultrastar-wingman/frontend/src/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "UltraStar Wingman",
        "version": "1.1.0"
    },
    "paths": {
...
        "/avatars/{avatar}": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Avatar",
                "operationId": "avatar_avatars__avatar__get",
                "parameters": [
                    {
                        "name": "avatar",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Avatar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/download": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Download",
                "operationId": "download_download_get",
                "parameters": [
                    {
                        "name": "view",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "View"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/usdb/download": {
            "post": {
                "tags": [
                    "USDB"
                ],
                "summary": "Downloads the song with the given USDB ID",
                "operationId": "api_usdb_download_api_usdb_download_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsdbId"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BasicResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usdb/songs": {
            "get": {
                "tags": [
                    "USDB"
                ],
                "summary": "Search Songs",
                "operationId": "api_usdb_songs_api_usdb_songs_get",
                "parameters": [
                    {
                        "name": "artist",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter songs by the artist's name.",
                            "title": "Artist"
                        },
                        "description": "Filter songs by the artist's name."
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter songs by title.",
                            "title": "Title"
                        },
                        "description": "Filter songs by title."
                    },
                    {
                        "name": "edition",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by the song's edition.",
                            "title": "Edition"
                        },
                        "description": "Filter by the song's edition."
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter songs by language.",
                            "title": "Language"
                        },
                        "description": "Filter songs by language."
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter songs by genre.",
                            "title": "Genre"
                        },
                        "description": "Filter songs by genre."
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/OrderEnum"
                                }
                            ],
                            "description": "Sort the result by this order criteria.",
                            "default": "rating",
                            "title": "Order"
                        },
                        "description": "Sort the result by this order criteria."
                    },
                    {
                        "name": "ud",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/UdEnum"
                                }
                            ],
                            "description": "Sort order: ascending (asc) or descending (desc).",
                            "default": "desc",
                            "title": "Ud"
                        },
                        "description": "Sort order: ascending (asc) or descending (desc)."
                    },
                    {
                        "name": "golden",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Golden"
                        }
                    },
                    {
                        "name": "songcheck",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Songcheck"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "The number of songs to return per page.",
                            "default": 30,
                            "title": "Limit"
                        },
                        "description": "The number of songs to return per page."
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "Page number for pagination.",
                            "default": 1,
                            "title": "Page"
                        },
                        "description": "Page number for pagination."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of songs matching the criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SongsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/songs/{song_id}/cover": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Cover",
                "operationId": "cover_api_songs__song_id__cover_get",
                "parameters": [
                    {
                        "name": "song_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Song Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Retrieve Players",
                "description": "Retrieves a list of all unique player names.",
                "operationId": "api_players_api_players_get",
                "responses": {
                    "200": {
                        "description": "A list of unique player names",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlayerList"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Add a New Player",
                "description": "Adds a new player name to the list.\n\n- **name**: The name of the player to add. It is taken from the form data.\n\nThis endpoint writes the new player's name to the players file, appending it to the end.\nIf the operation is successful, it returns a success message. Otherwise, it raises an HTTPException.",
                "operationId": "api_players_add_api_players_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlayerCreation"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Confirmation of player addition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlayerList"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/players/": {
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Delete a Player",
                "description": "Deletes a player name from the list.\n\n- **name**: The name of the player to delete.\n\nThis endpoint reads all player names, filters out the specified name, and rewrites the file without it.\nIf the operation is successful, it returns a success message.",
                "operationId": "delete_name_api_players__delete",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The name of the player to delete.",
                            "title": "Name"
                        },
                        "description": "The name of the player to delete."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation of player deletion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BasicResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/players/submit": {
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Submit Player Names",
                "description": "Submits a list of player names.\n\n- **names**: A list of names to be submitted.\n\nAccepts a list of names in the request body and submits them.",
                "operationId": "api_players_submit_api_players_submit_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlayerList"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Confirmation of successful name submission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BasicResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BasicResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success",
                        "description": "Success response",
                        "example": true
                    }
                },
                "type": "object",
                "title": "BasicResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "OrderEnum": {
                "type": "string",
                "enum": [
                    "id",
                    "artist",
                    "title",
                    "edition",
                    "rating",
                    "language",
                    "views",
                    "golden"
                ],
                "title": "OrderEnum"
            },
...
            "PlayerCreation": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "The player name."
                    }
                },
                "type": "object",
                "title": "PlayerCreation"
            },
            "PlayerList": {
                "properties": {
                    "players": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Players",
                        "description": "List of player names.",
                        "example": [
                            "Alice",
                            "Bob",
                            "Charlie"
                        ]
                    }
                },
                "type": "object",
                "title": "PlayerList"
            },
...
            "SongsResponse": {
                "properties": {
                    "paging": {
                        "$ref": "#/components/schemas/Paging"
                    },
                    "songs": {
                        "items": {
                            "$ref": "#/components/schemas/Song"
                        },
                        "type": "array",
                        "title": "Songs"
                    }
                },
                "type": "object",
                "required": [
                    "paging",
                    "songs"
                ],
                "title": "SongsResponse"
            },
            "UdEnum": {
                "type": "string",
                "enum": [
                    "asc",
                    "desc"
                ],
                "title": "UdEnum"
            },
            "UsdbId": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "The USDB ID."
                    }
                },
                "type": "object",
                "title": "UsdbId"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}